Different attempt at serializing using binary formatter directly
but this fails as Vector4 is not serializable
