+++ Thrive.csproj
@@ -190,7 +190,7 @@
     <Compile Include="src\saving\SaveContext.cs" />
     <Compile Include="src\saving\SaveHelper.cs" />
     <Compile Include="src\saving\SaveManager.cs" />
-    <Compile Include="src\saving\serializers\Base64ArrayConverter.cs" />
+    <Compile Include="src\saving\serializers\Base64BinaryConverter.cs" />
     <Compile Include="src\saving\serializers\BaseNodeConverter.cs" />
     <Compile Include="src\saving\serializers\CompoundCloudPlaneConverter.cs" />
     <Compile Include="src\saving\serializers\DynamicDeserializeObjectConverter.cs" />
